requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# HTTP Client
httpx>=0.25.0

# Data Processing
pandas>=2.1.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# External APIs
discogs-client>=2.3.0
pylast>=5.2.0

# Database (opcional, pero útil para migraciones futuras)
# psycopg2-binary>=2.9.0  # Descomentar si migras a PostgreSQL

# Utilities
streamlit>=1.28.0  # Para el explorador de DB